+++ changelog.md
@@ -122,7 +122,7 @@
 
 - `strformat.fmt` and `strformat.&` support `= specifier`. `fmt"{expr=}"` now expands to `fmt"expr={expr}"`.
 
-- Add `os.fromUnixPerm` convenience func to convert Unix-like file permission to `set[FilePermission]`,
+- Add `os.toFilePerm` convenience func to convert Unix-like file permission to `set[FilePermission]`,
   eg. you type `0o777` instead of `{fpUserExec, fpUserWrite, fpUserRead, fpGroupExec, fpGroupWrite, fpGroupRead, fpOthersExec, fpOthersWrite, fpOthersRead}`, etc.